+++ packages/openapi-formatter/src/commands.ts
@@ -7,13 +7,14 @@ import * as process from "node:process";
 export function createCommandFormat(program: Command) {
   program
     .command("format")
-    .description("Format Openapi specified files. Will mostly sort the document. For more formatting, refer the option.")
+    .description("Format Openapi specified files. Will by default sort the document. For more formatting, refer the option.")
     .argument("<openapi-spec>", "Absolut or Relative filepath from the cwd to the OpenApi root document file")
     .option("-o, --outputFolder [outputFolder]", "Absolut or relative filepath to the folder where all files will be written to")
-    .option("-a, --fix [fix]", "Apply all fixes", false)
+    .option("--sortSpec [sortSpec]", "Will sort the entire openapi spec", "true")
+    .option("-f, --fix [fix]", "Apply all fixes", false)
+    .option("--fixDanglingAllOfProps [fixDanglingAllOfProps]", "Insert sibling properties of an allOf within a schema as an allOf sub schema", false)
     .option("--fixDanglingAllOfProps [fixDanglingAllOfProps]", "Insert sibling properties of an allOf within a schema as an allOf sub schema", false)
     .option("--fixTitles [fixTitles]", "Insert or fix titles for schemas.", false)
-    .option("--fisDescription [fisDescription]", "Insert description for schemas", false)
     .option("-v, --verbose", "Enable verbose logging")
     .action(
       async (
@@ -22,7 +23,8 @@ export function createCommandFormat(program: Command) {
           outputFolder?: string;
           verbose: boolean;
           fix: boolean;
-        } & FormatterOptions
+          sortSpec: string
+        } & Omit<FormatterOptions, 'sortSpec'>
       ) => {
         if (options.verbose) {
           appLog.setLogLevel("debug");
@@ -32,7 +34,7 @@ export function createCommandFormat(program: Command) {
           appLog.log.error(`could not find specFile: ${specFile.absolutePath}`);
           process.exit(1);
         }
-
+        console.log("--options", JSON.stringify(options))
         const outFolder = options.outputFolder ? Folder.of(options.outputFolder) : specFile.folder;
         const result = await withPerformance(() =>
           formatSpec(specFile, {
@@ -40,6 +42,7 @@ export function createCommandFormat(program: Command) {
             fisDescription: options.fix || options.fisDescription,
             fixDanglingAllOfProps: options.fix || options.fixDanglingAllOfProps,
             fixTitles: options.fix || options.fixTitles,
+            sortSpec: options.fix || options.sortSpec === "true",
           })
         );
         appLog.log.info(`finished bundling in ${(result.duration / 1000).toFixed(3)} s`, { outFile: result.ret.outFile });