+++ packages/openapi-codegen-endpoints/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @dasaplan/openapi-codegen-endpoints
 
+## 0.0.12
+
+### Patch Changes
+
+- Updated dependencies
+  - @dasaplan/ts-sdk@0.0.2
+  - @dasaplan/openapi-bundler@0.0.20
+
 ## 0.0.11
 
 ### Patch Changes